# Optimize `URLHelpers.gather_constants`