@amomchilov 2024-07-12
I found one more trick which both simplifies the code (`.any? { |a| helper == a }` becomes just `.include?(helper)`), and speeds it up.

See the commit message of https://github.com/Shopify/tapioca/pull/1935/commits/ef5e6c9674fb77f59cd2f8a4f1003f83a442d73e for details.